Move getColumnExample into metabase-lib v1 expressions
Example values by column type are pure metadata-predicate logic with no
query-editing concern. Living in querying's CombineColumns util forced
ListViewConfiguration in visualizations to reach up into querying. It
now sits in metabase-lib/v1/expressions where both sides import
downward.
