+++ frontend/src/metabase/querying/components/expressions/CombineColumns/util.ts
@@ -2,6 +2,7 @@ import { t } from "ttag";
 
 import { isNotNull } from "metabase/utils/types";
 import * as Lib from "metabase-lib";
+import { getColumnExample } from "metabase-lib/v1/expressions/column-example";
 
 export type ColumnAndSeparator = {
   separator: string | null;
@@ -120,59 +121,6 @@ export const getExample = (
   columnsAndSeparators: ColumnAndSeparator[],
 ): string => {
   return flatten(columnsAndSeparators).map(getColumnExample).join("");
-};
-
-export const getColumnExample = (
-  column: Lib.ColumnMetadata | string | null,
-): string => {
-  if (!column) {
-    return "";
-  }
-  if (typeof column === "string") {
-    return column;
-  }
-
-  if (Lib.isEmail(column)) {
-    return "email@example.com";
-  }
-
-  if (Lib.isURL(column)) {
-    return "https://www.example.com";
-  }
-
-  if (Lib.isBoolean(column)) {
-    return "true";
-  }
-
-  if (Lib.isID(column)) {
-    return "12345";
-  }
-
-  if (Lib.isInteger(column)) {
-    return "123";
-  }
-
-  if (Lib.isNumeric(column)) {
-    return "123.45678901234567";
-  }
-
-  if (Lib.isDateWithoutTime(column)) {
-    return "2042-01-01";
-  }
-
-  if (Lib.isDateOrDateTime(column)) {
-    return "2042-01-01 12:34:56.789";
-  }
-
-  if (Lib.isTime(column)) {
-    return "12:34:56.789";
-  }
-
-  if (Lib.isLatitude(column) || Lib.isLongitude(column)) {
-    return "-12.34567";
-  }
-
-  return "text";
 };
 
 export function hasCombinations(columns: Lib.ColumnMetadata[]) {
