+++ frontend/src/metabase/visualizations/visualizations/List/components/ListView/ListViewConfiguration.tsx
@@ -16,7 +16,6 @@ import {
   ReorderableTagsInput,
   SortablePill,
 } from "metabase/common/components/ReorderableTagsInput/ReorderableTagsInput";
-import { getColumnExample } from "metabase/querying/components/expressions/CombineColumns/util";
 import {
   ActionIcon,
   Box,
@@ -34,6 +33,7 @@ import {
 import type { ColorName } from "metabase/ui/colors/types";
 import type { ComputedVisualizationSettings } from "metabase/viz-core";
 import type * as Lib from "metabase-lib";
+import { getColumnExample } from "metabase-lib/v1/expressions/column-example";
 import type {
   DatasetColumn,
   DatasetData,
